rename: 모달 컴포넌트 Gallery 폴더 내부로 이동
